Move Envmap and Skybox from 'scene/' to 'fx/'.
